All footballs in Sportmart
SportMart can be accessed anywhere by pressing q or the upper button on a gamepad. Still no currency system in place, so you basically pick whatever you want and get it for free.

+++ go-long-go-deep/Assets/Scripts/SkeletonController.cs
@@ -189,7 +189,7 @@ public class SkeletonController : MonoBehaviour {
 	 * else, play the taking damange animation and start a cooldown
 	 * to delay any further actions.
 	 */
-	void registerHit(int damage) {
+	void registerHit(float damage) {
 		gettingHit = true;
 //		Debug.Log ("Hit " + Time.fixedTime);
 		skeleBody.velocity = Vector2.zero;
